Adding support for text-shadow
