+++ packages/system/src/styles/index.js
@@ -6,6 +6,7 @@ import { layout } from './layout'
 import { flexboxes } from './flexboxes'
 import { grids } from './grids'
 import { positioning } from './positioning'
+import { shadows } from './shadows'
 import { space } from './space'
 import { typography } from './typography'
 import { xgrids } from './xgrids'
@@ -17,6 +18,7 @@ export * from './flexboxes'
 export * from './grids'
 export * from './layout'
 export * from './positioning'
+export * from './shadows'
 export * from './space'
 export * from './typography'
 export * from './xgrids'
@@ -29,6 +31,7 @@ export const system = compose(
   grids,
   layout,
   positioning,
+  shadows,
   space,
   typography,
   xgrids,
