+++ website/src/pages/docs/system-props.mdx
@@ -279,8 +279,14 @@ Available border utilities:
 - `borderWidth`
 - `borderStyle`
 - `borderRadius`
-- `boxShadow`
 - `borders`: all border utilities
+
+## Shadows
+
+Available shadow utilities:
+
+- `boxShadow`
+- `textShadow`
 
 ## Basics
